# osmpbf: use cgo/czlib for decompressions, with build tag