--- a/go.mod
+++ b/go.mod
@@ -3,6 +3,7 @@ module github.com/paulmach/osm
 go 1.13
 
 require (
+	github.com/datadog/czlib v0.0.0-20160811164712-4bc9a24e37f2
 	github.com/gogo/protobuf v1.3.1
 	github.com/paulmach/orb v0.1.6
 	github.com/paulmach/protoscan v0.1.0