--- a/osmpbf/zlib_cgo.go
+++ b/osmpbf/zlib_cgo.go
@@ -1,0 +1,14 @@
+// +build cgo
+
+package osmpbf
+
+import (
+	"bytes"
+	"io"
+
+	"github.com/datadog/czlib"
+)
+
+func zlibReader(data []byte) (io.ReadCloser, error) {
+	return czlib.NewReader(bytes.NewReader(data))
+}