--- a/osmpbf/zlib_go.go
+++ b/osmpbf/zlib_go.go
@@ -1,0 +1,13 @@
+// +build !cgo
+
+package osmpbf
+
+import (
+	"bytes"
+	"compress/zlib"
+	"io"
+)
+
+func zlibReader(data []byte) (io.ReadCloser, error) {
+	return zlib.NewReader(bytes.NewReader(data))
+}