The pbf data comes in blocks, each block is zlib compressed. Decompressing this data takes about 33% of the total read time on my benchmarks. I added support for a cgo version, https://github.com/DataDog/czlib with a build tag.

```
> CGO_ENABLED=0 go test -bench . 
BenchmarkLondon-12             3         341655178 ns/op        1245167664 B/op  4544349 allocs/op

> CGO_ENABLED=1 go test -bench . 
BenchmarkLondon-12             4         266967016 ns/op        1236313254 B/op  4492042 allocs/op

> benchcmp
benchmark              old ns/op     new ns/op     delta
BenchmarkLondon-12     341655178     266967016     -21.86%

benchmark              old allocs     new allocs     delta
BenchmarkLondon-12     4544349        4492042        -1.15%

benchmark              old bytes      new bytes      delta
BenchmarkLondon-12     1245167664     1236313254     -0.71%
```

The performance benefit is pretty large but I'm not sure what the community thinking around cgo is, so leaving this hear for now. I'm hoping the build tag support will make it usable.