@invisiblefunnel 2020-12-19
I think this is a good change. Another option is to let the caller override the zlib reader function. No strong preference.